moving Quiz to quiz dir
